# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "MovimientosMesResponse": {
        "additionalProperties": false,
        "description": "GET /api/movimientos — monthly transaction list (US-014).",
        "properties": {
          "periodo": {
            "description": "Resolved period, format YYYY-MM.",
            "type": "string"
          },
          "totalTransacciones": {
            "maximum": 9007199254740991,
            "minimum": -9007199254740991,
            "type": "integer"
          },
          "transacciones": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "abono": {
                  "description": "BigInt-safe decimal string amount (never a JSON number).",
                  "type": "string"
                },
                "banco": {
                  "type": "string"
                },
                "bucket": {
                  "description": "Bucket name (Bucket domain enum value).",
                  "type": "string"
                },
                "cargo": {
                  "description": "BigInt-safe decimal string amount (never a JSON number).",
                  "type": "string"
                },
                "categoria": {
                  "anyOf": [
                    {
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "nombre": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "nombre"
                      ],
                      "type": "object"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "description": "Folded category, or null for Ingreso rows or a row with no categoria assigned (issue #778 tramo 5b: SinCategoria no longer exists as a bucket)."
                },
                "descripcion": {
                  "type": "string"
                },
                "fecha": {
                  "description": "ISO-8601 UTC timestamp.",
                  "type": "string"
                },
                "id": {
                  "type": "string"
                },
                "numeroCuenta": {
                  "type": "string"
                },
                "tipoCuenta": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "fecha",
                "descripcion",
                "cargo",
                "abono",
                "banco",
                "tipoCuenta",
                "numeroCuenta",
                "bucket",
                "categoria"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "periodo",
          "totalTransacciones",
          "transacciones"
        ],
        "type": "object"
      },
...
      "RegistrarMovimientoManualResponse": {
        "additionalProperties": false,
        "description": "POST /api/movimientos 201 — the persisted manual movement (US-058, D-12). Amounts are BigInt-safe strings. `origen` is always \"Manual\".",
        "properties": {
          "abono": {
            "description": "BigInt-safe decimal string; \"0\" for Gasto rows.",
            "type": "string"
          },
          "bucket": {
            "description": "Bucket enum value (Ingreso | Necesidades | Deseos | Ahorro). Ingreso rows always carry \"Ingreso\"; Gasto rows carry the caller-supplied bucket.",
            "type": "string"
          },
          "cargo": {
            "description": "BigInt-safe decimal string; \"0\" for Ingreso rows.",
            "type": "string"
          },
          "categoriaId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "null for Ingreso rows; the caller-supplied categoriaId for Gasto rows."
          },
          "descripcion": {
            "type": "string"
          },
          "fecha": {
            "description": "ISO-8601 UTC timestamp.",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "origen": {
            "const": "Manual",
            "description": "Provenance marker — always \"Manual\" for this endpoint (C-a, D-13).",
            "type": "string"
          }
        },
        "required": [
          "id",
          "fecha",
          "descripcion",
          "cargo",
          "abono",
          "bucket",
          "categoriaId",
          "origen"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP contract for the MoneyDiary Express API, sourced from Zod schemas (ADR-011 amend).",
    "title": "MoneyDiary API",
    "version": "0.10.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ingestas/{id}": {
      "delete": {
        "description": "Authenticated endpoint that cascade-deletes an ingesta and its transactions (US-018, ING-01/ING-02). Requires x-api-key + a valid session (RNF-SEC-006, per-user isolation).",
        "parameters": [
          {
            "description": "Ingesta id (raw path param). Existence/ownership is checked by the use case, not this schema.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Ingesta id (raw path param). Existence/ownership is checked by the use case, not this schema.",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Ingesta deleted. No response body."
          },
          "404": {
            "description": "Anti-enumeration: the ingesta does not exist or does not belong to the authenticated user."
          }
        },
        "summary": "Delete an ingesta"
      }
    },
...
    "/api/movimientos": {
      "get": {
        "description": "Authenticated endpoint returning the consolidated monthly transaction list (US-014). Requires x-api-key + a valid session (RNF-SEC-006, per-user isolation).",
        "parameters": [
          {
            "description": "Month period, format YYYY-MM (e.g. 2026-07). Absent defaults to the latest month with at least one transaction for this user, falling back to the current month when the user has none (issue #747). Format is validated by the domain, not this schema.",
            "in": "query",
            "name": "periodo",
            "schema": {
              "description": "Month period, format YYYY-MM (e.g. 2026-07). Absent defaults to the latest month with at least one transaction for this user, falling back to the current month when the user has none (issue #747). Format is validated by the domain, not this schema.",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovimientosMesResponse"
                }
              }
            },
            "description": "Monthly transaction list for the resolved period."
          },
          "400": {
            "description": "Invalid periodo — either a transport-shape mismatch or a malformed YYYY-MM value (domain-level, PeriodoMes VO)."
          }
        },
        "summary": "Monthly transaction list"
      },
      "post": {
        "description": "Authenticated endpoint that registers a single hand-typed movement with no cartola (US-058, D-12). Type-first: `tipo=Ingreso` auto-classifies to {bucket=Ingreso, categoriaId=null}; `tipo=Gasto` requires `bucket` ∈ {Necesidades, Deseos, Ahorro} and a `categoriaId` from the caller's own catalog. The Ingreso variant is strict (stray `bucket`/`categoriaId` ⇒ 400, Q3 resolution). `monto` is a BigInt-safe decimal string (JSON number ⇒ 400). `fecha` must be YYYY-MM-DD; domain enforces ≤ today (future ⇒ 400). A per-user sentinel Account (banco='Manual') is created lazily (MAN-04). The persisted row has `ingestaId=null`, `origen='Manual'`, and is immune to `DELETE /api/ingestas/:id` (SQL NULL-equality, CA-04). Requires x-api-key + a valid session (RNF-SEC-006, per-user isolation, ISO-01/ISO-02).",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "additionalProperties": false,
                    "description": "Ingreso variant — strict (stray bucket/categoriaId ⇒ 400, Q3 resolution). Auto-classified to {bucket=Ingreso, categoriaId=null} by construction (D-10).",
                    "properties": {
                      "descripcion": {
                        "type": "string"
                      },
                      "fecha": {
                        "description": "Date of the movement, YYYY-MM-DD. Domain enforces ≤ today (future ⇒ 400).",
                        "type": "string"
                      },
                      "monto": {
                        "description": "BigInt-safe decimal string (never a JSON number). Domain enforces: positive integer, no float, overflow guard (> Number.MAX_SAFE_INTEGER ⇒ 400).",
                        "type": "string"
                      },
                      "tipo": {
                        "const": "Ingreso",
                        "type": "string"
                      }
                    },
                    "required": [
                      "tipo",
                      "fecha",
                      "descripcion",
                      "monto"
                    ],
                    "type": "object"
                  },
                  {
                    "description": "Gasto variant — requires bucket ∈ {Necesidades, Deseos, Ahorro} and categoriaId.",
                    "properties": {
                      "bucket": {
                        "description": "Required for Gasto. One of Necesidades | Deseos | Ahorro. Ingreso is invalid here (D-12).",
                        "enum": [
                          "Necesidades",
                          "Deseos",
                          "Ahorro"
                        ],
                        "type": "string"
                      },
                      "categoriaId": {
                        "description": "Required for Gasto. Must be a non-empty string belonging to the calling user catalog AND map to the supplied bucket (CategoriaFueraDeCatalogoError / BucketCategoriaNoConcuerdaError => 400).",
                        "minLength": 1,
                        "type": "string"
                      },
                      "descripcion": {
                        "type": "string"
                      },
                      "fecha": {
                        "description": "Date of the movement, YYYY-MM-DD.",
                        "type": "string"
                      },
                      "monto": {
                        "description": "BigInt-safe decimal string.",
                        "type": "string"
                      },
                      "tipo": {
                        "const": "Gasto",
                        "type": "string"
                      }
                    },
                    "required": [
                      "tipo",
                      "fecha",
                      "descripcion",
                      "monto",
                      "bucket",
                      "categoriaId"
                    ],
                    "type": "object"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrarMovimientoManualResponse"
                }
              }
            },
            "description": "Movement registered. Response carries id, fecha (ISO), descripcion (plaintext), cargo/abono (BigInt-safe strings), bucket, categoriaId, and origen=\"Manual\" (D-08/D-12). No DB read-back — values sourced from the in-memory VO."
          },
          "400": {
            "description": "Shape mismatch (Zod: stray fields on Ingreso, missing bucket/categoriaId on Gasto, monto as JSON number, wrong fecha format) OR domain validation failure (MovimientoManualInvalidoError: FECHA_FUTURA, DESCRIPCION_VACIA/LARGA, MONTO_INVALIDO/OVERFLOW, SIN_MONTOS, MONTO_NEGATIVO) OR catalog cascade failure (CategoriaFueraDeCatalogoError, BucketCategoriaNoConcuerdaError). Amounts are scrubbed from every error message (ADR-013)."
          },
          "500": {
            "description": "Infrastructure fault — sentinel upsert, catalog load, or Transaccion.create failed (PersistenciaFallidaError). Retryable."
          }
        },
        "summary": "Register a manual movement (type-first: Ingreso or Gasto)"
      }
    },
    "/api/movimientos/{id}": {
      "delete": {
        "description": "Authenticated endpoint that deletes a Transaccion owned by the calling user with origen='Manual' (correccion-movimientos-manuales, ADR-040). Ingesta-born rows are never individually deletable — delete the whole ingesta instead (DELETE /api/ingestas/:id). Requires x-api-key + a valid session (RNF-SEC-006, per-user isolation).",
        "parameters": [
          {
            "description": "Transaccion id (raw path param). Existence/ownership/provenance is checked by the use case, not this schema.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Transaccion id (raw path param). Existence/ownership/provenance is checked by the use case, not this schema.",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Movement deleted. No response body."
          },
          "404": {
            "description": "Anti-enumeration: the transaction does not exist, does not belong to the authenticated user, or is not origen='Manual' (ingesta-born rows collapse into this same response)."
          }
        },
        "summary": "Delete a manual movement"
      }
    },
...
  }
}